+++ examples/tour/Cargo.toml
@@ -8,26 +8,5 @@ repository = "https://github.com/hecrj/iced"
 edition = "2018"
 publish = false
 
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[[bin]]
-name = "main"
-path = "src/main.rs"
-
 [dependencies]
-futures-preview = "=0.3.0-alpha.18"
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-iced_native = { version = "0.1.0-alpha", path = "../../native" }
-# A personal `ggez` fork that introduces a `FontCache` type to measure text
-# efficiently and fixes HiDPI issues.
-ggez = { version = "0.5", git = "https://github.com/hecrj/ggez.git" }
-env_logger = "0.6"
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-iced_web = { path = "../../web" }
-wasm-bindgen = "0.2.50"
-log = "0.4"
-console_error_panic_hook = "0.1.6"
-console_log = "0.1.2"
+iced = { version = "0.1.0-alpha.1", path = "../.." }
