--- examples/tour/src/iced_ggez.rs
@@ -1,6 +1,0 @@
-mod renderer;
-mod widget;
-
-pub use renderer::Cache as ImageCache;
-pub use renderer::Renderer;
-pub use widget::*;
