--- examples/tour/src/widget.rs
@@ -1,5 +1,0 @@
-#[cfg(target_arch = "wasm32")]
-pub use iced_web::*;
-
-#[cfg(not(target_arch = "wasm32"))]
-pub use crate::iced_ggez::*;
