+++ winit/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "iced_winit"
+version = "0.1.0-alpha"
+authors = ["Héctor Ramón Jiménez <hector0193@gmail.com>"]
+edition = "2018"
+description = "A winit runtime for Iced"
+license = "MIT"
+repository = "https://github.com/hecrj/iced"
+
+[dependencies]
+iced_native = { version = "0.1.0-alpha", path = "../native" }
+winit = "0.20.0-alpha3"
